backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.5.2
sse-starlette>=1.8.2
azure-identity>=1.15.0
openai>=1.99.0
agent-framework-core
agent-framework-azure
requests>=2.31.0
